Readd Olex2 commands
