+++ services/applications/olex2_linux/configure_olex2.py
@@ -4,14 +4,14 @@ import shutil
 import subprocess
 from pathlib import Path
 
+from pyqcrbox import logger, sql_models
+from pyqcrbox.registry.client import QCrBoxClient
 from qcrboxtools.cif.cif2cif import cif_file_merge_to_unified_by_yml, cif_file_to_specific_by_yml
 from qcrboxtools.cif.merge import replace_structure_from_cif
 from qcrboxtools.robots.olex2 import Olex2Socket
 
-from pyqcrbox import logger, sql_models
-from pyqcrbox.registry.client import QCrBoxClient
-
 YAML_PATH = "./config_olex2.yaml"
+
 
 def run__interactive(input_file):
     logger.info("XXX DEBUG MODE: Running Olex2 via PythonCallable")
@@ -67,7 +67,6 @@ def prepare__interactive(input_file):
     pid_files = olex2_datadir.glob("*.olex2_pid")
     for pid_file in pid_files:
         pid_file.unlink()
-
 
 
 def finalise__interactive(input_file):
